Reset structure -> moved signal back into middleware.py
