+++ django_guid/__init__.py
@@ -1,7 +1,6 @@
 """
-If you have to put code here, make sure to test that you haven't broken setup.py
-You can test this by installing the package locally with `pip install -e .`
+Don't put imports or code here.
+This file is imported by setup.py
+Adding imports here will break setup.py
 """
-from . import signals  # noqa: F401
-
-__version__ = '1.2.0'
+__version__ = '1.1.1'
