+++ django_guid/middleware.py
@@ -2,12 +2,14 @@ import logging
 import threading
 import uuid
 from typing import Callable
+from typing import Optional
 from typing import Union
 
+from django.core.signals import request_finished
+from django.dispatch import receiver
 from django.http import HttpRequest, HttpResponse
 
-from django_guid.config import Settings
-
+from .config import settings
 
 logger = logging.getLogger('django_guid')
 
@@ -25,7 +27,6 @@ class GuidMiddleware(object):
 
     def __init__(self, get_response: Callable) -> None:
         self.get_response = get_response
-        self.settings = Settings()
 
     def __call__(self, request: HttpRequest) -> Union[HttpRequest, HttpResponse]:
         """
@@ -41,10 +42,10 @@ class GuidMiddleware(object):
 
         # ^ Code above this line is executed before the view and later middleware
         response = self.get_response(request)
-        if self.settings.RETURN_HEADER:
-            response[self.settings.GUID_HEADER_NAME] = self.get_guid()  # Adds the GUID to the response header
-            if self.settings.EXPOSE_HEADER:
-                response['Access-Control-Expose-Headers'] = self.settings.GUID_HEADER_NAME
+        if settings.RETURN_HEADER:
+            response[settings.GUID_HEADER_NAME] = self.get_guid()  # Adds the GUID to the response header
+            if settings.EXPOSE_HEADER:
+                response['Access-Control-Expose-Headers'] = settings.GUID_HEADER_NAME
         return response
 
     @classmethod
@@ -111,11 +112,11 @@ class GuidMiddleware(object):
         :param request: HttpRequest object
         :return: GUID
         """
-        given_guid = str(request.headers.get(self.settings.GUID_HEADER_NAME))
-        if not self.settings.VALIDATE_GUID:
+        given_guid = str(request.headers.get(settings.GUID_HEADER_NAME))
+        if not settings.VALIDATE_GUID:
             logger.debug('Returning ID from header without validating it as a GUID')
             return given_guid
-        elif self.settings.VALIDATE_GUID and self._validate_guid(given_guid):
+        elif settings.VALIDATE_GUID and self._validate_guid(given_guid):
             logger.debug('%s is a valid GUID', given_guid)
             return given_guid
         else:
@@ -132,7 +133,7 @@ class GuidMiddleware(object):
         :param request: HttpRequest object
         :return: GUID
         """
-        guid_header_name = self.settings.GUID_HEADER_NAME
+        guid_header_name = settings.GUID_HEADER_NAME
         header = request.headers.get(guid_header_name)  # Case insensitive headers.get added in Django2.2
         if header:
             logger.info('%s found in the header: %s', guid_header_name, header)
@@ -144,3 +145,17 @@ class GuidMiddleware(object):
             )
 
         return request.correlation_id
+
+
+@receiver(request_finished)
+def delete_guid(sender: Optional[dict], **kwargs: dict) -> None:
+    """
+    Receiver function for when a request finishes.
+    When a request is finished,
+    we make sure that the current requests _guid reference is deleted to prevent a memory leak.
+
+    :param sender: dict or None
+    :param kwargs: dict
+    :return: None
+    """
+    GuidMiddleware.delete_guid()
